0.0.3 Compatibility with webpack 3.x, optimize the code, bug fix
