+++ index.js
@@ -1,80 +1,20 @@
+/**
+ * Main file of file replace loader
+ * {@link https://github.com/vyushin/file-replace-loader/blob/master/index.js}
+ */
+
 const loaderUtils = require('loader-utils');
 const validateOptions = require('schema-utils');
 const path = require('path');
 const fs = require('fs');
+const CONSTANTS = require('./constants');
 
-/**
- * Usually used in read file functions
- * @const
- */
-const ENCODING = 'utf8';
-
-/**
- * Loader name
- * @const
- */
-const LOADER_NAME = 'file-replace-loader';
-
-/**
- * Loader replacement conditions
- * These modes have to use in loader options in webpack config.
- * Array indexing helps using this array like enum type during code is writing
- * @const
- * @enum
- */
-const LOADER_REPLACEMENT_CONDITIONS = [];
-LOADER_REPLACEMENT_CONDITIONS[0] = false; // Equivalent to "never"
-LOADER_REPLACEMENT_CONDITIONS[1] = true; // Equivalent to "always"
-LOADER_REPLACEMENT_CONDITIONS[2] = 'always';
-LOADER_REPLACEMENT_CONDITIONS[3] = 'never';
-LOADER_REPLACEMENT_CONDITIONS[4] = 'if-replacement-exists';
-LOADER_REPLACEMENT_CONDITIONS[5] = 'if-source-is-empty';
-
-/**
- * Schema for validate loader options
- * @const
- */
-const LOADER_OPTIONS_SCHEMA = {
-  type: 'object',
-  properties: {
-    condition: {
-      enum: LOADER_REPLACEMENT_CONDITIONS,
-      default: LOADER_REPLACEMENT_CONDITIONS[4],
-      errorMessages: {
-        enum: `should be equal to one of the allowed values: [${LOADER_REPLACEMENT_CONDITIONS.join(', ')}]`
-      }
-    },
-    replacement: {
-      type: 'string',
-    },
-    async: {
-      type: 'boolean',
-      default: true
-    }
-  },
-  additionalProperties: false,
-  required: ['replacement'],
-  getDefaultOptions() {
-    const result = {};
-    Object.keys(this.properties).forEach((key) => {
-      result[key] = this.properties[key].default;
-    });
-    return result;
-  },
-  getProperty(key) {
-    return this.properties[key];
-  }
-};
-
-/**
- * Error types that using in error messages
- * @const
- * @enum
- */
-const ERROR_TYPES = [];
-ERROR_TYPES[0] = 'Invalid options';
-ERROR_TYPES[1] = 'Replacement error';
-ERROR_TYPES[2] = 'File reading error';
+const ENCODING = CONSTANTS.ENCODING;
+const LOADER_NAME = CONSTANTS.LOADER_NAME;
+const LOADER_REPLACEMENT_CONDITIONS = CONSTANTS.LOADER_REPLACEMENT_CONDITIONS;
+const LOADER_OPTIONS_SCHEMA = CONSTANTS.LOADER_OPTIONS_SCHEMA;
+const ERROR_TYPES = CONSTANTS.ERROR_TYPES;
+const ERROR_MESSAGES = CONSTANTS.ERROR_MESSAGES;
 
 /**
  * Custom exception formatted to the loader format
@@ -107,27 +47,26 @@ function prepareErrorSchemaMessage(e) {
 }
 
 /**
- * Wrapped readFileSync
- * @param {String} path Path to file
- * @return {String} File content
+ * Formatting info messages
+ * @param {String} message
+ * @return {String}
  */
-function readFileSync(path) {
-  return fs.readFileSync(path, { encoding: ENCODING, flag: 'r' });
+function prepareInfoMessage(message) {
+  return `[${LOADER_NAME}] ${message}`;
 }
 
-/**
- * Wrapped readFileS
- * @param {String} path Path to file
- * @param {Function} callback
- */
-function readFile(path, callback) {
-  return fs.readFile(path, ENCODING, (err, content) => {
-    err && new Exception({
-      title: ERROR_TYPES[2],
-      message: err.message
+function readFile(path, isAsync, callback) {
+  if (isAsync) {
+    return fs.readFile(path, ENCODING, (err, content) => {
+      err && new Exception({
+        title: ERROR_TYPES[2],
+        message: err.message
+      });
+      callback(content);
     });
-    callback(content);
-  });
+  } else {
+    return fs.readFileSync(path, { encoding: ENCODING, flag: 'r' });
+  }
 }
 
 function getOptions(loaderContext) {
@@ -142,8 +81,8 @@ function getOptions(loaderContext) {
  */
 module.exports = function(source) {
   const options = getOptions(this);
-  const isSync = options && options.async === false;
-  const callback = isSync === false && this.async() || null;
+  const isAsync = options && options.async === true;
+  const callback = isAsync === true && this.async() || null;
 
   /**
    * Validate loader options before its work
@@ -161,20 +100,13 @@ module.exports = function(source) {
       options.condition === LOADER_REPLACEMENT_CONDITIONS[2]) {
     if (fs.existsSync(options.replacement)) {
       this.addDependency(options.replacement);
-      if (isSync) {
-        return readFileSync(options.replacement);
-      } else {
-        readFile(options.replacement, (content) => {
-          callback(null, content);
-        })
-      }
+      return isAsync
+        ? readFile(options.replacement, true, (content) => { callback(null, content) })
+        : readFile(options.replacement, false);
     } else {
       throw new Exception({
         title: ERROR_TYPES[1],
-        message: `File (${options.replacement}) doesn't exist but specified in ${LOADER_NAME} options with \n` +
-                 `  condition ${LOADER_REPLACEMENT_CONDITIONS[1]} or '${LOADER_REPLACEMENT_CONDITIONS[2]}'. \n` +
-                 `  Perhaps this is due replacement isn't full path. Make sure that file exists and replacement\n` +
-                 `  option is full path to file`
+        message: ERROR_MESSAGES[0].replace('$1', options.replacement),
       });
     }
   }
@@ -185,13 +117,9 @@ module.exports = function(source) {
   if (options.condition === LOADER_REPLACEMENT_CONDITIONS[4]) {
     if (fs.existsSync(options.replacement)) {
       this.addDependency(options.replacement);
-      if (isSync) {
-        return readFileSync(options.replacement)
-      } else {
-        readFile(options.replacement, (content) => {
-          callback(null, content);
-        })
-      }
+      return isAsync
+        ? readFile(options.replacement, true, (content) => { callback(null, content) })
+        : readFile(options.replacement, false);
     }
     /**
      * We don't need any errors here, because it isn't error when replacement doesn't exist by
@@ -203,29 +131,20 @@ module.exports = function(source) {
    * If condition is 'if-source-is-empty'
    */
   if (options.condition === LOADER_REPLACEMENT_CONDITIONS[5]) {
-    if (fs.existsSync(this.resourcePath)) {
+    if (fs.existsSync(options.replacement)) {
       const stats = fs.statSync(this.resourcePath);
       if (stats.size === 0) {
         this.addDependency(options.replacement);
-        if (isSync) {
-          return readFileSync(options.replacement)
-        } else {
-          readFile(options.replacement, (content) => {
-            callback(null, content);
-          })
-        }
+        return isAsync
+          ? readFile(options.replacement, true, (content) => { callback(null, content) })
+          : readFile(options.replacement, false);
       } else {
-        if (isSync) {
-          return source;
-        } else {
-          callback(null, source);
-        }
+        return isAsync ? callback(null, source) : source;
       }
     } else {
       throw new Exception({
         title: ERROR_TYPES[1],
-        message: `File (${this.resourcePath}) doesn't exist but specified in sources. ${LOADER_NAME} can't replace\n` +
-                 `  it to replacement file by '${LOADER_REPLACEMENT_CONDITIONS[5]}' condition. Make sure that source file exists.`
+        message: ERROR_MESSAGES[1].replace('$1', options.replacement),
       });
     }
   }
@@ -234,11 +153,7 @@ module.exports = function(source) {
    * If condition is 'never' or false
    */
   if (options.condition === LOADER_REPLACEMENT_CONDITIONS[0] ||
-    options.condition === LOADER_REPLACEMENT_CONDITIONS[3]) {
-    if (isSync) {
-      return source;
-    } else {
-      callback(null, source);
-    }
+      options.condition === LOADER_REPLACEMENT_CONDITIONS[3]) {
+    return isAsync ? callback(null, source) : source;
   }
 };
